Fix forward and reverse iterator when starting in middle of chunk (#57)
When starting in the middle of a chunk, its possible the messages for our
topic are _behind_ the position. In that case, we end up with an empty heap. This fixes the behavior to try again when the first set of candidate chunks results in no messages.

Add a test to ensure the behavior does not regress.