Slice underlying ArrayBuffer when reading a record (#67)
A more correct fix for avoiding individual messages holding on to the entire chunk memory.

The original attempt at this fix was in https://github.com/foxglove/rosbag/pull/62 but failed to account that the `slice` on a `Uint8Array` might still preserve the underlying ArrayBuffer array. This updated logic performs the slice on the underlying ArrayBuffer.